% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rappdirs-package.r
\docType{package}
\name{rappdirs-package}
\alias{rappdirs}
\alias{rappdirs-package}
\title{rappdirs: Application Directories: Determine Where to Save Data, Caches, and Logs}
\description{
An easy way to determine which directories on the users computer you should use to save data, caches and logs. A port of Python's 'Appdirs' (\url{https://github.com/ActiveState/appdirs}) to R.
}
\section{Main functions}{

\itemize{
\item \code{\link[=user_data_dir]{user_data_dir()}}
\item \code{\link[=user_config_dir]{user_config_dir()}}
\item \code{\link[=user_cache_dir]{user_cache_dir()}}
\item \code{\link[=site_data_dir]{site_data_dir()}}
\item \code{\link[=user_log_dir]{user_log_dir()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rappdirs.r-lib.org}
  \item \url{https://github.com/r-lib/rappdirs}
  \item Report bugs at \url{https://github.com/r-lib/rappdirs/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com} [translator, copyright holder]

Authors:
\itemize{
  \item Sridhar Ratnakumar
  \item Trent Mick
  \item Trevor Davis [translator]
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item ActiveState (R/appdir.r, R/cache.r, R/data.r, R/log.r translated from appdirs) [copyright holder]
  \item Eddy Petrisor [contributor]
  \item Gabor Csardi [contributor]
  \item Gregory Jefferis [contributor]
}

}
\keyword{internal}
